changed directory values from application.properties to be system independent changed interface methods of file producing services to be of top level
