+++ src/main/java/com/school/model/FileResource.java
@@ -2,6 +2,7 @@ package com.school.model;
 
 import com.school.configuration.FileConfig;
 
+import java.io.File;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
@@ -18,8 +19,8 @@ public abstract class FileResource {
         this.fileName = fileConfig.getName();
         this.fileDirectory = fileConfig.getDirectory();
         this.fileExtension = fileExtension;
-        //TODO:make it better
-        this.fullPathWithoutExtension = this.getFileDirectory() + fileConfig.getStudentsGradesSubdirectory() +
+        this.fullPathWithoutExtension = this.getFileDirectory() + File.separator +
+                fileConfig.getStudentsGradesSubdirectory() + File.separator +
                 fileConfig.getOptionalNamePrefix() +
                 LocalDateTime.now().format(this.getFormatter()) +
                 "_" +
