--- src/main/java/com/school/model/RecentlyProducedFile.kt
@@ -1,7 +1,0 @@
-package com.school.model
-
-import java.io.File
-
-interface RecentlyProducedFile {
-    fun getRecentlyProducedFile(directory: File?): FileToImport?
-}
