+++ src/main/java/com/school/service/fileproducers/ErrorCreatingFileException.kt
@@ -1,0 +1,3 @@
+package com.school.service.fileproducers
+
+class ErrorCreatingFileException(message: String): RuntimeException(message)
