+++ src/main/java/com/school/service/fileproducers/RecentlyProducedFile.kt
@@ -1,0 +1,9 @@
+package com.school.service.fileproducers
+
+import com.school.model.FileToImport
+import com.school.model.OptionalRequestParams
+
+interface RecentlyProducedFile {
+    fun importFile(params: OptionalRequestParams?): FileToImport
+    fun produceFileAndSaveToCorrespondingDirectory(directoryPath: String, params: OptionalRequestParams?)
+}
