+++ src/main/resources/application.properties
@@ -26,9 +26,9 @@ app.config.progress-records-directory=seed_performance/
 app.config.progress-records-file=performance_records.xls
 app.config.grade-topic=-grade-supplier
 
-config.file.directory=produced_files/
-config.file.students-grades-subdirectory=students_grades/
-config.file.students-list-subdirectory=students_list/
+config.file.directory=produced_files
+config.file.students-grades-subdirectory=students_grades
+config.file.students-list-subdirectory=students_list
 config.file.name=grades
 
 support-mail=baseschoolinformer@gmail.com
